feat(deposit): asset-first /deposit page (timeline + sidebar)
