--- src/routes/(authed)/(disabled)/deposit/+page.svelte
@@ -1,3 +1,0 @@
-<document:head>
-	<title>Deposit</title>
-</document:head>
